feat: extract search highlight logic into useSearchHighlight hook
Move result highlighting logic out of the 'GlobalSearch' component's inline anonymous functions and into a dedicated custom hook called 'useSearchHighlight'.

- Created `src/hooks/useSearchHighlight.tsx`
- Refactored `src/components/GlobalSearch.tsx` to use the new hook
- Cleaned up unused imports in `GlobalSearch.tsx`
